add serialization mod
